requirements.txt
flask>=3.0.0
tree-sitter==0.25.2
tree-sitter-python==0.25.0
pytest==8.3.4
pytest-cov==6.0.0
openai-agents==0.6.3
openai==2.13.0
pydantic==2.12.5
python-dotenv==1.2.1
rich>=13.0.0
